move IS3BucketOperations into metl-core (same location as IDirectory) (#972)
